Addressing code duplication
