+++ dev/AppNotifications/AppNotificationTelemetry.h
@@ -4,6 +4,7 @@
 #pragma once
 
 #include <WindowsAppRuntimeInsights.h>
+#include <NotificationTelemetryHelper.h>
 
 DECLARE_TRACELOGGING_CLASS(AppNotificationTelemetryProvider,
     "Microsoft.WindowsAppSDK.Notifications.AppNotificationTelemetry",
@@ -20,7 +21,7 @@ public:
         winrt::hresult hr,
         std::wstring const& appId) noexcept try
     {
-        if (c_maxEventLimit >= UpdateLogEventCount())
+        if (m_helper.ShouldLogEvent())
         {
             TraceLoggingClassWriteMeasure(
                 "Register",
@@ -28,8 +29,8 @@ public:
                 _GENERIC_PARTB_FIELDS_ENABLED,
                 TraceLoggingHexUInt32(hr, "OperationResult"),
                 TraceLoggingWideString(appId.c_str(), "AppId"),
-                TraceLoggingBool(IsPackagedApp(), "IsAppPackaged"),
-                TraceLoggingWideString(GetAppName().c_str(), "AppName"));
+                TraceLoggingBool(m_helper.IsPackagedApp(), "IsAppPackaged"),
+                TraceLoggingWideString(m_helper.GetAppName().c_str(), "AppName"));
         }
     }
     CATCH_LOG()
@@ -38,7 +39,7 @@ public:
         winrt::hresult hr,
         std::wstring const& appId) noexcept try
     {
-        if (c_maxEventLimit >= UpdateLogEventCount())
+        if (m_helper.ShouldLogEvent())
         {
             TraceLoggingClassWriteMeasure(
                 "Unregister",
@@ -46,8 +47,8 @@ public:
                 _GENERIC_PARTB_FIELDS_ENABLED,
                 TraceLoggingHexUInt32(hr, "OperationResult"),
                 TraceLoggingWideString(appId.c_str(), "AppId"),
-                TraceLoggingBool(IsPackagedApp(), "IsAppPackaged"),
-                TraceLoggingWideString(GetAppName().c_str(), "AppName"));
+                TraceLoggingBool(m_helper.IsPackagedApp(), "IsAppPackaged"),
+                TraceLoggingWideString(m_helper.GetAppName().c_str(), "AppName"));
         }
     }
     CATCH_LOG()
@@ -56,7 +57,7 @@ public:
         winrt::hresult hr,
         std::wstring const& appId) noexcept try
     {
-        if (c_maxEventLimit >= UpdateLogEventCount())
+        if (m_helper.ShouldLogEvent())
         {
             TraceLoggingClassWriteMeasure(
                 "UnregisterAll",
@@ -64,8 +65,8 @@ public:
                 _GENERIC_PARTB_FIELDS_ENABLED,
                 TraceLoggingHexUInt32(hr, "OperationResult"),
                 TraceLoggingWideString(appId.c_str(), "AppId"),
-                TraceLoggingBool(IsPackagedApp(), "IsAppPackaged"),
-                TraceLoggingWideString(GetAppName().c_str(), "AppName"));
+                TraceLoggingBool(m_helper.IsPackagedApp(), "IsAppPackaged"),
+                TraceLoggingWideString(m_helper.GetAppName().c_str(), "AppName"));
         }
     }
     CATCH_LOG()
@@ -77,7 +78,7 @@ public:
         winrt::hstring const& tag,
         winrt::hstring const& group) noexcept try
     {
-        if (c_maxEventLimit >= UpdateLogEventCount())
+        if (m_helper.ShouldLogEvent())
         {
             TraceLoggingClassWriteMeasure(
                 "Show",
@@ -88,8 +89,8 @@ public:
                 TraceLoggingUInt32(payload.size(), "PayloadSize"),
                 TraceLoggingUInt32(tag.size(), "TagSize"),
                 TraceLoggingUInt32(group.size(), "GroupSize"),
-                TraceLoggingBool(IsPackagedApp(), "IsAppPackaged"),
-                TraceLoggingWideString(GetAppName().c_str(), "AppName"));
+                TraceLoggingBool(m_helper.IsPackagedApp(), "IsAppPackaged"),
+                TraceLoggingWideString(m_helper.GetAppName().c_str(), "AppName"));
         }
     }
     CATCH_LOG()
@@ -100,7 +101,7 @@ public:
             winrt::hstring const& tag,
             winrt::hstring const& group) noexcept try
     {
-        if (c_maxEventLimit >= UpdateLogEventCount())
+        if (m_helper.ShouldLogEvent())
         {
             TraceLoggingClassWriteMeasure(
                 "UpdateAsync",
@@ -110,8 +111,8 @@ public:
                 TraceLoggingWideString(appId.c_str(), "AppId"),
                 TraceLoggingUInt32(tag.size(), "TagSize"),
                 TraceLoggingUInt32(group.size(), "GroupSize"),
-                TraceLoggingBool(IsPackagedApp(), "IsAppPackaged"),
-                TraceLoggingWideString(GetAppName().c_str(), "AppName"));
+                TraceLoggingBool(m_helper.IsPackagedApp(), "IsAppPackaged"),
+                TraceLoggingWideString(m_helper.GetAppName().c_str(), "AppName"));
         }
     }
     CATCH_LOG()
@@ -120,7 +121,7 @@ public:
         winrt::hresult hr,
         std::wstring const& appId) noexcept try
     {
-        if (c_maxEventLimit >= UpdateLogEventCount())
+        if (m_helper.ShouldLogEvent())
         {
             TraceLoggingClassWriteMeasure(
                 "Setting",
@@ -128,8 +129,8 @@ public:
                 _GENERIC_PARTB_FIELDS_ENABLED,
                 TraceLoggingHexUInt32(hr, "OperationResult"),
                 TraceLoggingWideString(appId.c_str(), "AppId"),
-                TraceLoggingBool(IsPackagedApp(), "IsAppPackaged"),
-                TraceLoggingWideString(GetAppName().c_str(), "AppName"));
+                TraceLoggingBool(m_helper.IsPackagedApp(), "IsAppPackaged"),
+                TraceLoggingWideString(m_helper.GetAppName().c_str(), "AppName"));
         }
     }
     CATCH_LOG()
@@ -139,7 +140,7 @@ public:
         std::wstring const& appId,
         uint32_t notificationId) noexcept try
     {
-        if (c_maxEventLimit >= UpdateLogEventCount())
+        if (m_helper.ShouldLogEvent())
         {
             TraceLoggingClassWriteMeasure(
                 "RemoveByIdAsync",
@@ -148,8 +149,8 @@ public:
                 TraceLoggingHexUInt32(hr, "OperationResult"),
                 TraceLoggingWideString(appId.c_str(), "AppId"),
                 TraceLoggingHexUInt32(notificationId, "NotificationId"),
-                TraceLoggingBool(IsPackagedApp(), "IsAppPackaged"),
-                TraceLoggingWideString(GetAppName().c_str(), "AppName"));
+                TraceLoggingBool(m_helper.IsPackagedApp(), "IsAppPackaged"),
+                TraceLoggingWideString(m_helper.GetAppName().c_str(), "AppName"));
         }
     }
     CATCH_LOG()
@@ -159,7 +160,7 @@ public:
         std::wstring const& appId,
         winrt::hstring const& tag) noexcept try
     {
-        if (c_maxEventLimit >= UpdateLogEventCount())
+        if (m_helper.ShouldLogEvent())
         {
             TraceLoggingClassWriteMeasure(
                 "RemoveByTagAsync",
@@ -168,8 +169,8 @@ public:
                 TraceLoggingHexUInt32(hr, "OperationResult"),
                 TraceLoggingWideString(appId.c_str(), "AppId"),
                 TraceLoggingUInt32(tag.size(), "TagSize"),
-                TraceLoggingBool(IsPackagedApp(), "IsAppPackaged"),
-                TraceLoggingWideString(GetAppName().c_str(), "AppName"));
+                TraceLoggingBool(m_helper.IsPackagedApp(), "IsAppPackaged"),
+                TraceLoggingWideString(m_helper.GetAppName().c_str(), "AppName"));
         }
     }
     CATCH_LOG()
@@ -180,7 +181,7 @@ public:
         winrt::hstring const& tag,
         winrt::hstring const& group) noexcept try
     {
-        if (c_maxEventLimit >= UpdateLogEventCount())
+        if (m_helper.ShouldLogEvent())
         {
             TraceLoggingClassWriteMeasure(
                 "RemoveByTagAndGroupAsync",
@@ -190,8 +191,8 @@ public:
                 TraceLoggingWideString(appId.c_str(), "AppId"),
                 TraceLoggingUInt32(tag.size(), "TagSize"),
                 TraceLoggingUInt32(group.size(), "GroupSize"),
-                TraceLoggingBool(IsPackagedApp(), "IsAppPackaged"),
-                TraceLoggingWideString(GetAppName().c_str(), "AppName"));
+                TraceLoggingBool(m_helper.IsPackagedApp(), "IsAppPackaged"),
+                TraceLoggingWideString(m_helper.GetAppName().c_str(), "AppName"));
         }
     }
     CATCH_LOG()
@@ -200,7 +201,7 @@ public:
         winrt::hresult hr,
         std::wstring const& appId) noexcept try
     {
-        if (c_maxEventLimit >= UpdateLogEventCount())
+        if (m_helper.ShouldLogEvent())
         {
             TraceLoggingClassWriteMeasure(
                 "RemoveByGroupAsync",
@@ -208,8 +209,8 @@ public:
                 _GENERIC_PARTB_FIELDS_ENABLED,
                 TraceLoggingHexUInt32(hr, "OperationResult"),
                 TraceLoggingWideString(appId.c_str(), "AppId"),
-                TraceLoggingBool(IsPackagedApp(), "IsAppPackaged"),
-                TraceLoggingWideString(GetAppName().c_str(), "AppName"));
+                TraceLoggingBool(m_helper.IsPackagedApp(), "IsAppPackaged"),
+                TraceLoggingWideString(m_helper.GetAppName().c_str(), "AppName"));
         }
     }
     CATCH_LOG()
@@ -218,7 +219,7 @@ public:
         winrt::hresult hr,
         std::wstring const& appId) noexcept try
     {
-        if (c_maxEventLimit >= UpdateLogEventCount())
+        if (m_helper.ShouldLogEvent())
         {
             TraceLoggingClassWriteMeasure(
                 "RemoveAllAsync",
@@ -226,8 +227,8 @@ public:
                 _GENERIC_PARTB_FIELDS_ENABLED,
                 TraceLoggingHexUInt32(hr, "OperationResult"),
                 TraceLoggingWideString(appId.c_str(), "AppId"),
-                TraceLoggingBool(IsPackagedApp(), "IsAppPackaged"),
-                TraceLoggingWideString(GetAppName().c_str(), "AppName"));
+                TraceLoggingBool(m_helper.IsPackagedApp(), "IsAppPackaged"),
+                TraceLoggingWideString(m_helper.GetAppName().c_str(), "AppName"));
         }
     }
     CATCH_LOG()
@@ -236,7 +237,7 @@ public:
         winrt::hresult hr,
         std::wstring const& appId) noexcept try
     {
-        if (c_maxEventLimit >= UpdateLogEventCount())
+        if (m_helper.ShouldLogEvent())
         {
             TraceLoggingClassWriteMeasure(
                 "GetAllAsync",
@@ -244,74 +245,12 @@ public:
                 _GENERIC_PARTB_FIELDS_ENABLED,
                 TraceLoggingHexUInt32(hr, "OperationResult"),
                 TraceLoggingWideString(appId.c_str(), "AppId"),
-                TraceLoggingBool(IsPackagedApp(), "IsAppPackaged"),
-                TraceLoggingWideString(GetAppName().c_str(), "AppName"));
+                TraceLoggingBool(m_helper.IsPackagedApp(), "IsAppPackaged"),
+                TraceLoggingWideString(m_helper.GetAppName().c_str(), "AppName"));
         }
     }
     CATCH_LOG()
 
 private:
-    wil::srwlock m_lock;
-    ULONGLONG m_lastFiredTick = 0;
-    UINT m_eventCount = 0;
-
-    static constexpr ULONGLONG c_logPeriod = 1000; // One second
-    static constexpr UINT c_maxEventLimit = 10;
-
-    UINT UpdateLogEventCount()
-    {
-        ULONGLONG currentTick = GetTickCount64();
-
-        auto lock{ m_lock.lock_exclusive() };
-
-        // Only fire limiting events every log period to prevent too many events from being fired
-        if ((currentTick - m_lastFiredTick) > c_logPeriod)
-        {
-            m_eventCount = 0;
-            m_lastFiredTick = currentTick;
-        }
-
-        m_eventCount++;
-
-        return m_eventCount;
-    }
-
-    inline bool IsPackagedApp() const
-    {
-        static const bool isPackagedApp{ AppModel::Identity::IsPackagedProcess() };
-
-        return isPackagedApp;
-    }
-
-    inline const std::wstring& GetAppName() const
-    {
-        static const std::wstring appName{ IsPackagedApp() ? GetAppNamePackaged() : GetAppNameUnpackaged() };
-
-        return appName;
-    }
-
-    std::wstring GetAppNamePackaged() const
-    {
-        wchar_t appUserModelId[APPLICATION_USER_MODEL_ID_MAX_LENGTH]{};
-
-        UINT32 appUserModelIdSize{ ARRAYSIZE(appUserModelId) };
-        THROW_IF_WIN32_ERROR(GetCurrentApplicationUserModelId(&appUserModelIdSize, appUserModelId));
-
-        return appUserModelId;
-    }
-
-    std::wstring CensorFilePath(const std::wstring& filepath) const
-    {
-        return { !PathIsFileSpecW(filepath.c_str()) ? PathFindFileNameW(filepath.c_str()) : filepath };
-    }
-
-    std::wstring GetAppNameUnpackaged() const
-    {
-        std::wstring appName;
-
-        wil::unique_cotaskmem_string processName;
-        THROW_IF_FAILED(wil::GetModuleFileNameExW(GetCurrentProcess(), nullptr, processName));
-
-        return CensorFilePath(processName.get());
-    }
+    NotificationTelemetryHelper m_helper;
 };
