+++ dev/Common/Common.vcxitems
@@ -24,6 +24,7 @@
     <ClInclude Include="$(MSBuildThisFileDirectory)AppModel.PackageGraph.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)Microsoft.Foundation.String.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)Microsoft.Utf8.h" />
+    <ClInclude Include="$(MSBuildThisFileDirectory)NotificationTelemetryHelper.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)Security.IntegrityLevel.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)WindowsAppRuntime.SelfContained.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)WindowsAppRuntime.VersionInfo.h" />
