+++ dev/PushNotifications/PushNotificationTelemetry.h
@@ -4,6 +4,7 @@
 #pragma once
 
 #include <WindowsAppRuntimeInsights.h>
+#include <NotificationTelemetryHelper.h>
 
 DECLARE_TRACELOGGING_CLASS(PushNotificationTelemetryProvider,
     "Microsoft.WindowsAppSDK.Notifications.PushNotificationTelemetry",
@@ -20,7 +21,7 @@ public:
         winrt::hresult hr,
         const winrt::guid& remoteId) noexcept try
     {
-        if (c_maxEventLimit >= UpdateLogEventCount())
+        if (m_helper.ShouldLogEvent())
         {
             TraceLoggingClassWriteMeasure(
                 "CreateChannelAsync",
@@ -28,8 +29,8 @@ public:
                 _GENERIC_PARTB_FIELDS_ENABLED,
                 TraceLoggingHexUInt32(hr, "OperationResult"),
                 TraceLoggingGuid(remoteId, "RemoteId"),
-                TraceLoggingBool(IsPackagedApp(), "IsAppPackaged"),
-                TraceLoggingWideString(GetAppName().c_str(), "AppName"));
+                TraceLoggingBool(m_helper.IsPackagedApp(), "IsAppPackaged"),
+                TraceLoggingWideString(m_helper.GetAppName().c_str(), "AppName"));
         }
     }
     CATCH_LOG()
@@ -37,15 +38,15 @@ public:
     DEFINE_EVENT_METHOD(LogCloseChannel)(
         winrt::hresult hr) noexcept try
     {
-        if (c_maxEventLimit >= UpdateLogEventCount())
+        if (m_helper.ShouldLogEvent())
         {
             TraceLoggingClassWriteMeasure(
                 "CloseChannel",
                 TelemetryPrivacyDataTag(PDT_ProductAndServicePerformance),
                 _GENERIC_PARTB_FIELDS_ENABLED,
                 TraceLoggingHexUInt32(hr, "OperationResult"),
-                TraceLoggingBool(IsPackagedApp(), "IsAppPackaged"),
-                TraceLoggingWideString(GetAppName().c_str(), "AppName"));
+                TraceLoggingBool(m_helper.IsPackagedApp(), "IsAppPackaged"),
+                TraceLoggingWideString(m_helper.GetAppName().c_str(), "AppName"));
         }
     }
     CATCH_LOG()
@@ -53,15 +54,15 @@ public:
     DEFINE_EVENT_METHOD(LogRegister)(
         winrt::hresult hr) noexcept try
     {
-        if (c_maxEventLimit >= UpdateLogEventCount())
+        if (m_helper.ShouldLogEvent())
         {
             TraceLoggingClassWriteMeasure(
                 "Register",
                 TelemetryPrivacyDataTag(PDT_ProductAndServicePerformance), 
                 _GENERIC_PARTB_FIELDS_ENABLED,
                 TraceLoggingHexUInt32(hr, "OperationResult"),
-                TraceLoggingBool(IsPackagedApp(), "IsAppPackaged"),
-                TraceLoggingWideString(GetAppName().c_str(), "AppName"));
+                TraceLoggingBool(m_helper.IsPackagedApp(), "IsAppPackaged"),
+                TraceLoggingWideString(m_helper.GetAppName().c_str(), "AppName"));
         }
     }
     CATCH_LOG()
@@ -69,15 +70,15 @@ public:
     DEFINE_EVENT_METHOD(LogUnregister)(
         winrt::hresult hr) noexcept try
     {
-        if (c_maxEventLimit >= UpdateLogEventCount())
+        if (m_helper.ShouldLogEvent())
         {
             TraceLoggingClassWriteMeasure(
                 "Unregister",
                 TelemetryPrivacyDataTag(PDT_ProductAndServicePerformance),
                 _GENERIC_PARTB_FIELDS_ENABLED,
                 TraceLoggingHexUInt32(hr, "OperationResult"),
-                TraceLoggingBool(IsPackagedApp(), "IsAppPackaged"),
-                TraceLoggingWideString(GetAppName().c_str(), "AppName"));
+                TraceLoggingBool(m_helper.IsPackagedApp(), "IsAppPackaged"),
+                TraceLoggingWideString(m_helper.GetAppName().c_str(), "AppName"));
         }
     }
     CATCH_LOG()
@@ -85,81 +86,20 @@ public:
     DEFINE_EVENT_METHOD(LogUnregisterAll)(
         winrt::hresult hr) noexcept try
     {
-        if (c_maxEventLimit >= UpdateLogEventCount())
+        if (m_helper.ShouldLogEvent())
         {
             TraceLoggingClassWriteMeasure(
                 "UnregisterAll",
                 TelemetryPrivacyDataTag(PDT_ProductAndServicePerformance),
                 _GENERIC_PARTB_FIELDS_ENABLED,
                 TraceLoggingHexUInt32(hr, "OperationResult"),
-                TraceLoggingBool(IsPackagedApp(), "IsAppPackaged"),
-                TraceLoggingWideString(GetAppName().c_str(), "AppName"));
+                TraceLoggingBool(m_helper.IsPackagedApp(), "IsAppPackaged"),
+                TraceLoggingWideString(m_helper.GetAppName().c_str(), "AppName"));
         }
     }
     CATCH_LOG()
 
 private:
-    wil::srwlock m_lock;
-    ULONGLONG m_lastFiredTick = 0;
-    UINT m_eventCount = 0;
+    NotificationTelemetryHelper m_helper;
 
-    static constexpr ULONGLONG c_logPeriod = 1000; // One second
-    static constexpr UINT c_maxEventLimit = 10;
-
-    UINT UpdateLogEventCount()
-    {
-        ULONGLONG currentTick = GetTickCount64();
-
-        auto lock{ m_lock.lock_exclusive() };
-
-        // Only fire limiting events every log period to prevent too many events from being fired
-        if ((currentTick - m_lastFiredTick) > c_logPeriod)
-        {
-            m_eventCount = 0;
-            m_lastFiredTick = currentTick;
-        }
-
-        m_eventCount++;
-
-        return m_eventCount;
-    }
-
-    inline bool IsPackagedApp() const
-    {
-        static const bool isPackagedApp{ AppModel::Identity::IsPackagedProcess() };
-
-        return isPackagedApp;
-    }
-
-    inline const std::wstring& GetAppName() const
-    {
-        static const std::wstring appName{ IsPackagedApp() ? GetAppNamePackaged() : GetAppNameUnpackaged() };
-
-        return appName;
-    }
-
-    std::wstring GetAppNamePackaged() const
-    {
-        wchar_t appUserModelId[APPLICATION_USER_MODEL_ID_MAX_LENGTH]{};
-
-        UINT32 appUserModelIdSize{ ARRAYSIZE(appUserModelId) };
-        THROW_IF_WIN32_ERROR(GetCurrentApplicationUserModelId(&appUserModelIdSize, appUserModelId));
-
-        return appUserModelId;
-    }
-
-    std::wstring CensorFilePath(const std::wstring& filepath) const
-    {
-        return { !PathIsFileSpecW(filepath.c_str()) ? PathFindFileNameW(filepath.c_str()) : filepath };
-    }
-
-    std::wstring GetAppNameUnpackaged() const
-    {
-        std::wstring appName;
-
-        wil::unique_cotaskmem_string processName;
-        THROW_IF_FAILED(wil::GetModuleFileNameExW(GetCurrentProcess(), nullptr, processName));
-
-        return CensorFilePath(processName.get());
-    }
 };
